+++ app/Http/Controllers/Admin/TrainerController.php
@@ -20,9 +20,13 @@ class TrainerController extends Controller
         return view('admin.trainers.index')->with($data);
     }
 
-    public function create()
+    public function create($id=null)
     {
-        return view('admin.trainers.create');
+        if($id!=null)
+        {
+            $data['trainer']=Trainer::findOrFail($id);
+        }
+        return view('admin.trainers.create')->with($data);
     }
 
     public function store(Request $request)
@@ -31,48 +35,44 @@ class TrainerController extends Controller
             'name'=>'required|string',
             'spec'=>'required|string',
             'phone'=>'nullable|string',
-            'img'=>'required|image|mimes:jpg,jpeg,png'
-        ]);
-        $new_name=$data['img']->hashName();
 
-        Image::make($data['img'])->resize(50,50)->save(public_path('uploads/trainers/'.$new_name));
-        $data['img']=$new_name;
-        Trainer::create($data);
+        ]);
+
+
+        if($request->id!=null)
+        {
+            $data+=$request->validate(['img'=>'nullable|image|mimes:jpg,jpeg,png']);
+            $old_name=Trainer::findOrFail($request->id)->img;
+            if($request->hasFile('img'))
+            {
+
+                Storage::disk('uploads')->delete('trainers/'.$old_name);
+                $new_name=$data['img']->hashName();
+                Image::make($data['img'])->resize(50,50)->save(public_path('uploads/trainers/'.$new_name));
+                $data['img']=$new_name;
+
+            }
+            else{
+                $data['img']=$old_name;
+            }
+
+            Trainer::findOrFail($request->id)->update($data);
+        }else{
+            $data=$request->validate(['img'=>'required|image|mimes:jpg,jpeg,png']);
+            $new_name=$data['img']->hashName();
+            Image::make($data['img'])->resize(50,50)->save(public_path('uploads/trainers/'.$new_name));
+            $data['img']=$new_name;
+            Trainer::create($data);
+
+        }
 
         return redirect(route('admin.trainers.index'));
 
     }
 
-    public function edit($id)
-    {
-        $data['trainer']=Trainer::findOrFail($id);
-        return view('admin.trainers.edit')->with($data);
-    }
-
     public function update(Request $request)
     {
-        $data=$request->validate([
-            'name'=>'required|string',
-            'spec'=>'required|string',
-            'phone'=>'nullable|string',
-            'img'=>'nullable|image|mimes:jpg,jpeg,png'
-        ]);
 
-        $old_name=Trainer::findOrFail($request->id)->img;
-        if($request->hasFile('img'))
-        {
-
-            Storage::disk('uploads')->delete('trainers/'.$old_name);
-            $new_name=$data['img']->hashName();
-            Image::make($data['img'])->resize(50,50)->save(public_path('uploads/trainers/'.$new_name));
-            $data['img']=$new_name;
-
-        }
-        else{
-            $data['img']=$old_name;
-        }
-
-        Trainer::findOrFail($request->id)->update($data);
         return back();
     }
